NewGameChooser - Move factory method for a NewGameChooser Comparator to be in NewGameChooser.java from NewGameChooserModel.java
